Add a test for verifying Key-Value table put operation with auth on

+++ test/integration/src/main/java/io/pravega/test/integration/demo/ClusterWrapper.java
@@ -176,7 +176,7 @@ public class ClusterWrapper implements AutoCloseable {
             segmentStoreServer.close();
             serviceBuilder.close();
             zookeeperServer.close();
-            if (passwordInputFile.exists()) {
+            if (passwordInputFile != null && passwordInputFile.exists()) {
                 passwordInputFile.delete();
             }
         } catch (Exception e) {
